Upgraded Model Instantiation and removed redundant code (#8)
* Reconfigured model instantiation

* Refactor download_usd.py to download and extract a specific zip file from Google Drive

* Add gdown to dependencies in setup.py
