+++ rover_envs/envs/__init__.py
@@ -2,6 +2,6 @@ import os
 
 # import gymnasium as gym
 
-ORBIT_CUSTOM_ENVS_EXT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../"))
+ROVER_LAB_CUSTOM_ENVS_EXT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../"))
 
-ORBIT_CUSTOM_ENVS_DATA_DIR = os.path.join(ORBIT_CUSTOM_ENVS_EXT_DIR, "learning")
+ROVER_LAB_ENVS_DATA_DIR = os.path.join(ROVER_LAB_CUSTOM_ENVS_EXT_DIR, "envs", "navigation", "learning")
