+++ rover_envs/envs/navigation/learning/skrl/configure_models.py
@@ -2,10 +2,11 @@ from gymnasium.spaces.box import Box
 from isaaclab.envs import ManagerBasedRLEnv
 
 from rover_envs.envs.navigation.learning.skrl.models import (Critic, DeterministicActor, DeterministicNeuralNetwork,
-                                                             GaussianNeuralNetwork)
+                                                             DeterministicNeuralNetworkConv, GaussianNeuralNetwork,
+                                                             GaussianNeuralNetworkConv)
 
 
-def get_models(agent: str, env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
+def get_models(agent: str, env: ManagerBasedRLEnv, observation_space: Box, action_space: Box, conv: bool = False):
     """
     Placeholder function for getting the models.
 
@@ -20,20 +21,22 @@ def get_models(agent: str, env: ManagerBasedRLEnv, observation_space: Box, actio
     """
 
     if agent == "PPO":
-        return gaussian_model_skrl(env, observation_space, action_space)
+        if conv:
+            return get_model_gaussian_conv(env, observation_space, action_space)
+        return get_model_gaussian(env, observation_space, action_space)
     if agent == "TRPO":
-        return gaussian_model_skrl(env, observation_space, action_space)
+        return get_model_gaussian(env, observation_space, action_space)
     if agent == "RPO":
-        return gaussian_model_skrl(env, observation_space, action_space)
+        return get_model_gaussian(env, observation_space, action_space)
     if agent == "SAC":
-        return double_critic_deterministic_model_skrl(env, observation_space, action_space)
+        return get_model_double_critic_deterministic(env, observation_space, action_space)
     if agent == "TD3":
-        return double_critic_deterministic_model_skrl(env, observation_space, action_space)
+        return get_model_double_critic_deterministic(env, observation_space, action_space)
 
     raise ValueError(f"Agent {agent} not supported.")
 
 
-def gaussian_model_skrl(env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
+def get_model_gaussian(env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
     models = {}
     encoder_input_size = env.unwrapped.observation_manager.group_obs_term_dim["policy"][-1][0]
 
@@ -42,7 +45,7 @@ def gaussian_model_skrl(env: ManagerBasedRLEnv, observation_space: Box, action_s
     models["policy"] = GaussianNeuralNetwork(
         observation_space=observation_space,
         action_space=action_space,
-        device=env.device,
+        device=env.unwrapped.device,
         mlp_input_size=mlp_input_size,
         mlp_layers=[256, 160, 128],
         mlp_activation="leaky_relu",
@@ -53,7 +56,7 @@ def gaussian_model_skrl(env: ManagerBasedRLEnv, observation_space: Box, action_s
     models["value"] = DeterministicNeuralNetwork(
         observation_space=observation_space,
         action_space=action_space,
-        device=env.device,
+        device=env.unwrapped.device,
         mlp_input_size=mlp_input_size,
         mlp_layers=[256, 160, 128],
         mlp_activation="leaky_relu",
@@ -64,7 +67,38 @@ def gaussian_model_skrl(env: ManagerBasedRLEnv, observation_space: Box, action_s
     return models
 
 
-def double_critic_deterministic_model_skrl(env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
+def get_model_gaussian_conv(env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
+    models = {}
+    encoder_input_size = env.unwrapped.observation_manager.group_obs_term_dim["policy"][-1][0]
+
+    mlp_input_size = 5
+
+    models["policy"] = GaussianNeuralNetworkConv(
+        observation_space=observation_space,
+        action_space=action_space,
+        device=env.device,
+        mlp_input_size=mlp_input_size,
+        mlp_layers=[256, 160, 128],
+        mlp_activation="leaky_relu",
+        encoder_input_size=encoder_input_size,
+        encoder_layers=[8, 16, 32, 64],
+        encoder_activation="leaky_relu",
+    )
+    models["value"] = DeterministicNeuralNetworkConv(
+        observation_space=observation_space,
+        action_space=action_space,
+        device=env.device,
+        mlp_input_size=mlp_input_size,
+        mlp_layers=[256, 160, 128],
+        mlp_activation="leaky_relu",
+        encoder_input_size=encoder_input_size,
+        encoder_layers=[8, 16, 32, 64],
+        encoder_activation="leaky_relu",
+    )
+    return models
+
+
+def get_model_double_critic_deterministic(env: ManagerBasedRLEnv, observation_space: Box, action_space: Box):
     models = {}
     encoder_input_size = env.unwrapped.observation_manager.group_obs_term_dim["policy"][-1][0]
