+++ rover_envs/envs/navigation/robots/aau_rover/__init__.py
@@ -11,7 +11,7 @@ gym.register(
     disable_env_checker=True,
     kwargs={
         "env_cfg_entry_point": env_cfg.AAURoverEnvCfg,
-        "best_model_path": f"{os.path.dirname(__file__)}/policies/best_agent.pt",
+        "best_model_path": f"{os.path.dirname(__file__)}/policies/best_agent2.pt",
         "get_agent_fn": get_agent,
     }
 )
