+++ rover_envs/utils/config.py
@@ -18,28 +18,28 @@ try:
 except ImportError:
     from skrl.utils.model_instantiators.torch import Shape  # noqa: F401
 
-from rover_envs.envs import ORBIT_CUSTOM_ENVS_DATA_DIR
+from rover_envs.envs import ROVER_LAB_ENVS_DATA_DIR
 
 __all__ = ["SKRL_PPO_CONFIG_FILE", "parse_skrl_cfg"]
 
 
 SKRL_PPO_CONFIG_FILE = {
     # classic
-    "AAURoverEnv-v0": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "AAURoverEnvCamera-v0": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "AAURoverEnvNew-v0": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "AAURoverEnvNoObstacles-v0": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "AAURoverEnv-v0_PPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "AAURoverEnv-v0_TRPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_trpo.yaml"),
-    "AAURoverEnv-v0_TD3": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_td3.yaml"),
-    "AAURoverEnv-v0_SAC": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_sac.yaml"),
-    "AAURoverEnv-v0_RPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_rpo.yaml"),
-    "Exomy-v0_PPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
-    "Exomy-v0_TRPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_trpo.yaml"),
-    "Exomy-v0_TD3": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_td3.yaml"),
-    "Exomy-v0_SAC": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_sac.yaml"),
-    "Exomy-v0_RPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_rpo.yaml"),
-    "AAURoverEnvCamera-v0_PPO": os.path.join(ORBIT_CUSTOM_ENVS_DATA_DIR, "skrl/rover_ppo.yaml"),
+    "AAURoverEnv-v0": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "AAURoverEnvCamera-v0": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "AAURoverEnvNew-v0": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "AAURoverEnvNoObstacles-v0": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "AAURoverEnv-v0_PPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "AAURoverEnv-v0_TRPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_trpo.yaml"),
+    "AAURoverEnv-v0_TD3": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_td3.yaml"),
+    "AAURoverEnv-v0_SAC": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_sac.yaml"),
+    "AAURoverEnv-v0_RPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_rpo.yaml"),
+    "Exomy-v0_PPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
+    "Exomy-v0_TRPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_trpo.yaml"),
+    "Exomy-v0_TD3": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_td3.yaml"),
+    "Exomy-v0_SAC": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_sac.yaml"),
+    "Exomy-v0_RPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_rpo.yaml"),
+    "AAURoverEnvCamera-v0_PPO": os.path.join(ROVER_LAB_ENVS_DATA_DIR, "skrl/configs/rover_ppo.yaml"),
 }
